feat: move common pdf processing functions to shared submissions.util
